requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask==1.1.2
graphviz==0.8.4
idna==2.6
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.2
mxnet-cu101==1.5.0
numpy==1.16.6
opencv-python==4.4.0.44
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.18.4
six==1.15.0
urllib3==1.22
Werkzeug==1.0.1
wincertstore==0.2
